+++ library-src/Sound/Punctual/WebGL.hs
@@ -38,6 +38,7 @@ import Sound.Punctual.GL
 import Sound.Punctual.AsyncProgram
 import Sound.Punctual.Resolution
 import Sound.Punctual.Texture
+import qualified Sound.Punctual.Webcam as Webcam
 
 data PunctualWebGL = PunctualWebGL {
   theCanvas :: HTMLCanvasElement,
@@ -46,8 +47,7 @@ data PunctualWebGL = PunctualWebGL {
   triangleStrip :: WebGLBuffer,
   fftTexture :: WebGLTexture,
   ifftTexture :: WebGLTexture,
-  camVideo :: HTMLVideoElement,
-  camTexture :: WebGLTexture,
+  webcam :: Webcam.Webcam,
   textures :: Map TextureRef Texture,
   fb0 :: (WebGLFramebuffer,WebGLTexture),
   fb1 :: (WebGLFramebuffer,WebGLTexture),
@@ -157,30 +157,11 @@ getAudioOutputAnalysis glCtx st = case (audioOutputAnalyser st) of
 
 foreign import javascript safe
   "$1.bindTexture($1.TEXTURE_2D, $3);\
-  \$1.texImage2D($1.TEXTURE_2D, 0, $1.LUMINANCE, 64, 1, 0, $1.LUMINANCE, $1.UNSIGNED_BYTE, $2);\
+  \$1.texImage2D($1.TEXTURE_2D, 0, $1.LUMINANCE, 512, 1, 0, $1.LUMINANCE, $1.UNSIGNED_BYTE, $2);\
   \$1.texParameteri($1.TEXTURE_2D, $1.TEXTURE_WRAP_S, $1.CLAMP_TO_EDGE);\
   \$1.texParameteri($1.TEXTURE_2D, $1.TEXTURE_WRAP_T, $1.CLAMP_TO_EDGE);\
   \$1.texParameteri($1.TEXTURE_2D, $1.TEXTURE_MIN_FILTER, $1.LINEAR);"
   _fftToTexture :: WebGLRenderingContext -> JSVal -> WebGLTexture -> IO ()
-
-
--- experimental webcam-as-texture support
-
-foreign import javascript safe
-  "$r = document.createElement('video');\
-  \$r.width = 2048; $r.height = 2048; $r.autoplay = true; $r.isPlaying = false;\
-  \$r.addEventListener('playing',function() { $r.isPlaying = true; });\
-  \navigator.mediaDevices.getUserMedia({video: true}).then( function(stream) { $r.srcObject = stream; } );"
-  videoForWebCam :: IO HTMLVideoElement
-
-foreign import javascript safe
-  "if($3.isPlaying) { $1.activeTexture($1.TEXTURE3);\
-  \$1.bindTexture($1.TEXTURE_2D,$2);\
-  \$1.texImage2D($1.TEXTURE_2D, 0, $1.RGBA, $1.RGBA, $1.UNSIGNED_BYTE, $3);\
-  \$1.texParameteri($1.TEXTURE_2D, $1.TEXTURE_WRAP_S, $1.CLAMP_TO_EDGE);\
-  \$1.texParameteri($1.TEXTURE_2D, $1.TEXTURE_WRAP_T, $1.CLAMP_TO_EDGE);\
-  \$1.texParameteri($1.TEXTURE_2D, $1.TEXTURE_MIN_FILTER, $1.LINEAR); } "
-  updateWebCamTexture :: WebGLRenderingContext -> WebGLTexture -> HTMLVideoElement -> IO ()
 
 
 -- given a list of requested texture sources (images) and the previous map of requested texture sources
@@ -204,12 +185,10 @@ newPunctualWebGL mic out res _brightness cvs ctx = runGL ctx $ do
   ts <- createBuffer
   bindBufferArray ts
   liftIO $ bufferDataArrayStatic glCtx
-  -- create textures to hold output fft (FFT), input fft (IFFT), and webcam data
+  -- create textures to hold output fft (FFT), input fft (IFFT)
   oFFT <- createTexture
   iFFT <- createTexture
-  -- allocate permanent objects for webcam-as-texture access
-  camVid <- liftIO $ videoForWebCam
-  camTex <- createTexture
+  wc <- Webcam.new
   -- create two framebuffers to ping-pong between as backbuffer/feedback
   frameBuffer0 <- makeFrameBufferTexture res
   frameBuffer1 <- makeFrameBufferTexture res
@@ -222,8 +201,7 @@ newPunctualWebGL mic out res _brightness cvs ctx = runGL ctx $ do
     triangleStrip = ts,
     fftTexture = oFFT,
     ifftTexture = iFFT,
-    camVideo = camVid,
-    camTexture = camTex,
+    webcam = wc,
     textures = Map.empty,
     fb0 = frameBuffer0,
     fb1 = frameBuffer1,
@@ -340,9 +318,7 @@ evaluatePunctualWebGL ctx tempo z p st = runGL ctx $ do
 
   let prevAsync = IntMap.findWithDefault emptyAsyncProgram z (mainPrograms st)
   newAsync <- updateAsyncProgram prevAsync defaultVertexShader newFragmentShader
-  let prevProgramsNeedAudioInputAnalysis = elem True $ fmap programNeedsAudioInputAnalysis $ IntMap.elems $ newPrevPrograms
-  let prevProgramsNeedAudioOutputAnalysis = elem True $ fmap programNeedsAudioOutputAnalysis $ IntMap.elems $ newPrevPrograms
-
+  let currAndPrevPrograms = IntMap.elems newCurrPrograms ++ IntMap.elems newPrevPrograms
   let st' = st {
     prevPrograms = newPrevPrograms,
     currPrograms = newCurrPrograms,
@@ -350,21 +326,24 @@ evaluatePunctualWebGL ctx tempo z p st = runGL ctx $ do
     textures = newTextures,
     textureMapsEval = newTextureMaps,
     firstZone = head $ IntMap.keys newCurrPrograms, -- recalculate which zone is first in drawing order so that defaultAlpha can be correct
-    needsAudioInputAnalysis = programNeedsAudioInputAnalysis p || prevProgramsNeedAudioInputAnalysis,
-    needsAudioOutputAnalysis = programNeedsAudioOutputAnalysis p || prevProgramsNeedAudioOutputAnalysis
+    needsAudioInputAnalysis = elem True $ fmap programNeedsAudioInputAnalysis currAndPrevPrograms,
+    needsAudioOutputAnalysis =elem True $ fmap programNeedsAudioOutputAnalysis currAndPrevPrograms
     }
   st'' <- liftIO $ updateAudioAnalysis st'
+  setWebcamActive st''
   return (st'',newFragmentShader)
+  
+  
+setWebcamActive :: PunctualWebGL -> GL ()
+setWebcamActive st = do
+  -- the webcam should be active if any current or previous programs in any zone require it, otherwise it should be inactive
+  let shouldBeActive = (elem True $ fmap programNeedsWebcam $ IntMap.elems $ currPrograms st) || (elem True $ fmap programNeedsWebcam $ IntMap.elems $ prevPrograms st)
+  Webcam.setActive (webcam st) shouldBeActive
 
-updateWebCam :: PunctualWebGL -> GL ()
-updateWebCam st = do
-  rCtx <- gl
-  -- liftIO $ updateWebCamImage (camVideo st) (camCanvas st) (camImage st)
-  liftIO $ updateWebCamTexture rCtx (camTexture st) (camVideo st)
 
 drawPunctualWebGL :: GLContext -> Tempo -> UTCTime -> Int -> PunctualWebGL -> IO PunctualWebGL
 drawPunctualWebGL ctx tempo now z st = runGL ctx $ do
-  updateWebCam st
+  Webcam.updateTexture (webcam st)
   let mainUniforms = ["res","width","height","_fb","_cam","_fft","_ifft","tex0","tex1","tex2","tex3","tex4","tex5","tex6","tex7","tex8","tex9","tex10","tex11","tex12","lo","mid","hi","ilo","imid","ihi","_defaultAlpha","_cps","_time","_etime","_beat","_ebeat"]
   let mainAttribs = ["p"]
   let prevAsync = IntMap.findWithDefault emptyAsyncProgram z $ mainPrograms st
@@ -392,7 +371,7 @@ drawPunctualWebGL ctx tempo now z st = runGL ctx $ do
     bindTex 1 (fftTexture st') fftLoc
     bindTex 2 (ifftTexture st') ifftLoc
     camLoc <- getUniformLocation program "_cam"
-    bindTex 3 (camTexture st') camLoc
+    bindTex 3 (Webcam._texture $ webcam st') camLoc
     -- bind textures to uniforms representing textures in the program
     let uMap = uniformsMap asyncProgram
     let texs = IntMap.findWithDefault (Map.empty) z $ textureMapsDraw st' -- Map Text Int